Add js to check if passwords are the same
Adds javascript file to check whether or not the two password fields
have the same input that differs from empty input. If they are the same,
make '*passwords do not match' text hidden and remove invalid classes
from the inputs. If they are not the same, change the visibility
attribute to visible and add invalid classes for the inputs.